add Product

+++ src/main/java/com/thebest/thebestpc/service/cart/CartServiceImpl.java
@@ -8,6 +8,7 @@ import com.thebest.thebestpc.repository.CartRepository;
 import com.thebest.thebestpc.repository.ProductRepository;
 import com.thebest.thebestpc.repository.UsersRepository;
 import com.thebest.thebestpc.service.cartItem.CartItemService;
+import com.thebest.thebestpc.service.cartItem.CartItemServiceImpl;
 import com.thebest.thebestpc.service.users.UsersService;
 
 import org.springframework.stereotype.Service;
@@ -19,17 +20,15 @@ public class CartServiceImpl implements CartService {
 
     private final UsersRepository usersRepository;
     private final CartRepository cartRepository;
-    private final ProductRepository productRepository;
-    private final CartItemService cartItemService;
     private final UsersService usersService;
+    private final CartItemServiceImpl cartItemServiceImpl;
 
 
-    public CartServiceImpl(UsersRepository usersRepository, CartRepository cartRepository, ProductRepository productRepository, CartItemService cartService, UsersService usersService) {
+    public CartServiceImpl(UsersRepository usersRepository, CartRepository cartRepository, UsersService usersService, CartItemServiceImpl cartItemServiceImpl) {
         this.usersRepository = usersRepository;
         this.cartRepository = cartRepository;
-        this.productRepository = productRepository;
-        this.cartItemService = cartService;
         this.usersService = usersService;
+        this.cartItemServiceImpl = cartItemServiceImpl;
     }
 
     @Override
@@ -37,17 +36,20 @@ public class CartServiceImpl implements CartService {
         if (!this.checkout(userId)) {
             Users users = usersRepository.findById(userId).orElseThrow(() -> new RuntimeException("User not found"));
             Cart cart = Cart.builder().users(users).build();
-            cartRepository.save(cart);
-            users.setCart(cart);
+            if (users.getCart() == null) {
+                cartRepository.save(cart);
+                users.setCart(cart);
+            }
             usersService.updateSecurityContext(users);
         }
-
     }
 
     @Override
     @Transactional
     public void removeCart(String userId) {
+        cartItemServiceImpl.getCartItems("1");
         Cart cart = findByUsersId(userId);
+        if (cart == null) throw new IllegalArgumentException("Tài khoản chua có giỏ hàng không thể xóa ");
         cartRepository.delete(cart);
     }
 
@@ -61,16 +63,10 @@ public class CartServiceImpl implements CartService {
         return cartRepository.findByUsersId(userId).isPresent();
     }
 
-    @Override
-    public void addCartItemToCart(String userId, Long productId) {
-        Cart cart = cartRepository.findByUsersId(userId).orElseThrow(() -> new RuntimeException("Cart not found"));
-        Product product = productRepository.findById(productId).orElseThrow(() -> new RuntimeException("Product not found"));
-        cartItemService.addCartItem(cart, product);
-    }
 
     @Override
     public Cart getCart(String userId) {
-        return cartRepository.findByUsersId(userId).orElseThrow(() -> new RuntimeException("Cart not found"));
+        return cartRepository.findByUsersId(userId).orElse(null);
     }
 
     @Override
